+++ vectorize.h
@@ -1,4 +1,4 @@
 #include <stdint.h>
 
-float gcc_dot_product(float *x, float *y, unsigned int n);
+double gcc_dot_product(float *x, float *y, unsigned int n);
